// src/api/models/school.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sanchalak Mobile API
 * API documentation for Sanchalak Mobile Application (Parent/Student)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@sanchalak.in
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ContactInfo } from './contact-info';

export interface School {
    'createdAt'?: string;
    'updatedAt'?: string;
    'id'?: string;
    'schoolCode': string;
    'name': string;
    'registrationNumber'?: string;
    'status'?: SchoolStatusEnum;
    'timezone'?: string;
    'board'?: string;
    'contactInfo'?: ContactInfo;
}

export const SchoolStatusEnum = {
    Draft: 'DRAFT',
    Active: 'ACTIVE',
    PaymentDue: 'PAYMENT_DUE',
    Restricted: 'RESTRICTED',
    Suspended: 'SUSPENDED',
    Archived: 'ARCHIVED'
} as const;

export type SchoolStatusEnum = typeof SchoolStatusEnum[keyof typeof SchoolStatusEnum];